Version 0.1.860031
Fixed crash when changing permissions while the service is running.

Fixed MANAGE_UNKNOWN_APP_SOURCES not showing on Android 12.

The app will now download its updates and ask you to allow it to install them or to approve an update.

Also, the addon will now try to start the service again if it's not responding.