Version 0.1.860037
Added missing query.

Fixed crash when changing permissions while the service is running caused by the system restarting the service with null intent. Also, if this happens while updating the update will be restarted but only if the version is still different.

Fixed MANAGE_UNKNOWN_APP_SOURCES not showing on Android 12.

The app will now download its updates and ask you to allow it to install them or to approve an update.

Also, the addon will now try to start the service again if it's not responding.